requirements.txt
Flask==2.3.2
Flask-Security-Too
flask_sqlalchemy
Jinja2==3.1.2
mysqlclient
flask-socketio
python-dotenv==1.0.0
requests
Werkzeug==2.3.3
bcrypt==4.0.1
Gunicorn